Fix words/memes selection in captcha
The current captcha system uses a random number between 0 and 3 to determine whether or not to use a string from the Words array or from the Memes array for each part of the captcha. However, all numbers in that range evaluate to true in Ruby. Therefore, only the Words array is used. This change fixes that, giving each part of the captcha a 25% chance to be selected from the Memes array.